WriteBlock(...,len) changed to uint at -r1.56 of libmythtv/util.cpp
git-svn-id: http://svn.mythtv.org/svn/trunk@6523 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythbackend/filetransfer.cpp
@@ -79,8 +79,8 @@ int FileTransfer::RequestBlock(int size)
         
         if (rbuffer->GetStopReads() || ret <= 0)
             break;
-            
-        if (!WriteBlock(sock->socketDevice(), requestBuffer, ret))
+
+        if (!WriteBlock(sock->socketDevice(), requestBuffer, (uint)ret))
         {
             tot = -1;
             break;
